requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
statsmodels==0.14.0
flask==3.0.0
scikit-learn==1.3.0
pytest==7.4.3
python-dateutil==2.8.2